split add-trailing-comma into plugins
