+++ add_trailing_comma/_main.py
@@ -1,563 +1,24 @@
 import argparse
-import ast
-import collections
-import functools
 import sys
-import warnings
-from typing import Callable
-from typing import Dict
-from typing import Generator
 from typing import Iterable
 from typing import List
-from typing import NamedTuple
 from typing import Optional
 from typing import Sequence
-from typing import Set
 from typing import Tuple
-from typing import Type
-from typing import TYPE_CHECKING
-from typing import TypeVar
 
-from tokenize_rt import ESCAPED_NL
-from tokenize_rt import NON_CODING_TOKENS
-from tokenize_rt import Offset
 from tokenize_rt import src_to_tokens
 from tokenize_rt import Token
 from tokenize_rt import tokens_to_src
-from tokenize_rt import UNIMPORTANT_WS
 
-if TYPE_CHECKING:
-    from typing import Protocol
-else:
-    Protocol = object
+from add_trailing_comma._ast_helpers import ast_parse
+from add_trailing_comma._data import FUNCS
+from add_trailing_comma._data import visit
+from add_trailing_comma._token_helpers import find_simple
+from add_trailing_comma._token_helpers import fix_brace
+from add_trailing_comma._token_helpers import START_BRACES
 
-NEWLINES = frozenset((ESCAPED_NL, 'NEWLINE', 'NL'))
-INDENT_TOKENS = frozenset(('INDENT', UNIMPORTANT_WS))
-START_BRACES = frozenset(('(', '{', '['))
-END_BRACES = frozenset((')', '}', ']'))
 
-
-class ParseState(NamedTuple):
-    in_fstring: bool = False
-
-
-AST_T = TypeVar('AST_T', bound=ast.AST)
-TokenFunc = Callable[[int, List[Token], Tuple[int, ...]], None]
-ASTFunc = Callable[[ParseState, AST_T], Iterable[Tuple[Offset, TokenFunc]]]
-
-
-class ASTCallbackMapping(Protocol):
-    def __getitem__(self, tp: Type[AST_T]) -> List[ASTFunc[AST_T]]: ...
-
-
-class Fix(NamedTuple):
-    braces: Tuple[int, int]
-    multi_arg: bool
-    remove_comma: bool
-    initial_indent: int
-
-
-def ast_parse(contents_text: str) -> ast.Module:
-    with warnings.catch_warnings():
-        warnings.simplefilter('ignore')
-        return ast.parse(contents_text.encode())
-
-
-def _to_offset(node: ast.AST) -> Offset:
-    candidates = [node]
-    while candidates:
-        candidate = candidates.pop()
-        if hasattr(candidate, 'lineno'):
-            return Offset(candidate.lineno, candidate.col_offset)
-        elif hasattr(candidate, '_fields'):  # pragma: no cover (<py39)
-            for field in reversed(candidate._fields):
-                candidates.append(getattr(candidate, field))
-    else:
-        raise AssertionError(node)
-
-
-def visit(
-        funcs: ASTCallbackMapping,
-        tree: ast.AST,
-) -> Dict[Offset, List[TokenFunc]]:
-    nodes = [(tree, ParseState())]
-
-    ret = collections.defaultdict(list)
-    while nodes:
-        node, parse_state = nodes.pop()
-
-        tp = type(node)
-        for ast_func in funcs[tp]:
-            for offset, token_func in ast_func(parse_state, node):
-                ret[offset].append(token_func)
-
-        if tp is ast.FormattedValue:
-            parse_state = parse_state._replace(in_fstring=True)
-
-        for name in reversed(node._fields):
-            value = getattr(node, name)
-            if isinstance(value, ast.AST):
-                nodes.append((value, parse_state))
-            elif isinstance(value, list):
-                for value in reversed(value):
-                    if isinstance(value, ast.AST):
-                        nodes.append((value, parse_state))
-    return ret
-
-
-FUNCS = collections.defaultdict(list)
-
-
-def register(
-        tp: Type[AST_T],
-) -> Callable[[ASTFunc[AST_T]], ASTFunc[AST_T]]:
-    def register_decorator(func: ASTFunc[AST_T]) -> ASTFunc[AST_T]:
-        FUNCS[tp].append(func)
-        return func
-    return register_decorator
-
-
-def _fix_literal(
-        i: int,
-        tokens: List[Token],
-        version: Tuple[int, ...],
-        *,
-        one_el_tuple: bool,
-) -> None:
-    if tokens[i].src in START_BRACES:  # pragma: no branch (<py38)
-        _fix_brace(
-            tokens, _find_simple(i, tokens),
-            add_comma=True,
-            remove_comma=not one_el_tuple,
-        )
-
-
-@register(ast.Set)
-def visit_Set(
-        parse_state: ParseState,
-        node: ast.Set,
-) -> Iterable[Tuple[Offset, TokenFunc]]:
-    func = functools.partial(_fix_literal, one_el_tuple=False)
-    yield _to_offset(node), func
-
-
-@register(ast.List)
-def visit_List(
-        parse_state: ParseState,
-        node: ast.List,
-) -> Iterable[Tuple[Offset, TokenFunc]]:
-    if node.elts:
-        func = functools.partial(_fix_literal, one_el_tuple=False)
-        yield _to_offset(node), func
-
-
-@register(ast.Dict)
-def visit_Dict(
-        parse_state: ParseState,
-        node: ast.Dict,
-) -> Iterable[Tuple[Offset, TokenFunc]]:
-    if node.values:
-        func = functools.partial(_fix_literal, one_el_tuple=False)
-        yield _to_offset(node), func
-
-
-def _fix_tuple(
-        i: int,
-        tokens: List[Token],
-        version: Tuple[int, ...],
-        *,
-        one_el_tuple: bool,
-) -> None:
-    _fix_brace(
-        tokens,
-        _find_tuple(i, tokens),
-        add_comma=True,
-        remove_comma=not one_el_tuple,
-    )
-
-
-@register(ast.Tuple)
-def visit_Tuple(
-        parse_state: ParseState,
-        node: ast.Tuple,
-) -> Iterable[Tuple[Offset, TokenFunc]]:
-    if node.elts:
-        is_one_el = _one_el_tuple(node)
-        if (
-                _to_offset(node) == _to_offset(node.elts[0]) or
-                # in < py38 tuples lie about offset -- later we must backtrack
-                sys.version_info < (3, 8)
-        ):
-            func = functools.partial(_fix_tuple, one_el_tuple=is_one_el)
-            yield _to_offset(node), func
-        else:  # pragma: no cover (py38+)
-            func = functools.partial(_fix_literal, one_el_tuple=is_one_el)
-            yield _to_offset(node), func
-
-
-def _fix_call(
-        i: int,
-        tokens: List[Token],
-        version: Tuple[int, ...],
-        *,
-        starargs: bool,
-        arg_offsets: Set[Offset],
-) -> None:
-    return _fix_brace(
-        tokens,
-        _find_call(arg_offsets, i, tokens),
-        add_comma=not starargs or version >= (3, 5),
-        remove_comma=True,
-    )
-
-
-@register(ast.Call)
-def visit_Call(
-        parse_state: ParseState,
-        node: ast.Call,
-) -> Iterable[Tuple[Offset, TokenFunc]]:
-    argnodes = [*node.args, *node.keywords]
-    arg_offsets = set()
-    has_starargs = False
-    for argnode in argnodes:
-        if isinstance(argnode, ast.Starred):
-            has_starargs = True
-        if isinstance(argnode, ast.keyword) and argnode.arg is None:
-            has_starargs = True
-
-        offset = _to_offset(argnode)
-        # multiline strings have invalid position, ignore them
-        if offset.utf8_byte_offset != -1:  # pragma: no branch (cpy bug)
-            arg_offsets.add(offset)
-
-    # If the sole argument is a generator, don't add a trailing comma as
-    # this breaks lib2to3 based tools
-    only_a_generator = (
-        len(argnodes) == 1 and isinstance(argnodes[0], ast.GeneratorExp)
-    )
-
-    if arg_offsets and not only_a_generator and not parse_state.in_fstring:
-        func = functools.partial(
-            _fix_call,
-            starargs=has_starargs,
-            arg_offsets=arg_offsets,
-        )
-        yield _to_offset(node), func
-
-
-def _fix_func(
-        i: int,
-        tokens: List[Token],
-        version: Tuple[int, ...],
-        *,
-        starargs: bool,
-        arg_offsets: Set[Offset],
-) -> None:
-    _fix_brace(
-        tokens,
-        _find_call(arg_offsets, i, tokens),
-        add_comma=not starargs or version >= (3, 6),
-        remove_comma=True,
-    )
-
-
-@register(ast.FunctionDef)
-def visit_FunctionDef(
-        parse_state: ParseState,
-        node: ast.FunctionDef,
-) -> Iterable[Tuple[Offset, TokenFunc]]:
-    has_starargs = False
-    args = [*getattr(node.args, 'posonlyargs', ()), *node.args.args]
-
-    if node.args.vararg:
-        args.append(node.args.vararg)
-        has_starargs = True
-    if node.args.kwarg:
-        args.append(node.args.kwarg)
-        has_starargs = True
-    if node.args.kwonlyargs:
-        args.extend(node.args.kwonlyargs)
-        has_starargs = True
-
-    arg_offsets = {_to_offset(arg) for arg in args}
-
-    if arg_offsets:
-        func = functools.partial(
-            _fix_func,
-            starargs=has_starargs,
-            arg_offsets=arg_offsets,
-        )
-        yield _to_offset(node), func
-
-
-def _fix_import(i: int, tokens: List[Token], version: Tuple[int, ...]) -> None:
-    _fix_brace(
-        tokens,
-        _find_import(i, tokens),
-        add_comma=True,
-        remove_comma=True,
-    )
-
-
-@register(ast.ImportFrom)
-def visit_ImportFrom(
-        parse_state: ParseState,
-        node: ast.ImportFrom,
-) -> Iterable[Tuple[Offset, TokenFunc]]:
-    yield _to_offset(node), _fix_import
-
-
-def _fix_class(
-        i: int,
-        tokens: List[Token],
-        version: Tuple[int, ...],
-        *,
-        arg_offsets: Set[Offset],
-) -> None:
-    _fix_brace(
-        tokens,
-        _find_call(arg_offsets, i, tokens),
-        add_comma=True,
-        remove_comma=True,
-    )
-
-
-@register(ast.ClassDef)
-def visit_ClassDef(
-        parse_state: ParseState,
-        node: ast.ClassDef,
-) -> Iterable[Tuple[Offset, TokenFunc]]:
-    # starargs are allowed in py3 class definitions, py35+ allows trailing
-    # commas.  py34 does not, but adding an option for this very obscure
-    # case seems not worth it.
-    args = [*node.bases, *node.keywords]
-    arg_offsets = {_to_offset(arg) for arg in args}
-
-    if arg_offsets:
-        func = functools.partial(_fix_class, arg_offsets=arg_offsets)
-        yield _to_offset(node), func
-
-
-def _find_simple(first_brace: int, tokens: List[Token]) -> Optional[Fix]:
-    brace_stack = [first_brace]
-    multi_arg = False
-
-    for i in range(first_brace + 1, len(tokens)):
-        token = tokens[i]
-        if token.src in START_BRACES:
-            brace_stack.append(i)
-        elif token.src in END_BRACES:
-            brace_stack.pop()
-
-        if len(brace_stack) == 1 and token.src == ',':
-            multi_arg = True
-
-        if not brace_stack:
-            break
-    else:
-        raise AssertionError('Past end?')
-
-    last_brace = i
-
-    # Check if we're actually multi-line
-    if (
-            # we were single line, but with an extra comma and or whitespace
-            tokens[first_brace].line == tokens[last_brace].line and (
-                tokens[last_brace - 1].name == UNIMPORTANT_WS or
-                tokens[last_brace - 1].src == ','
-            )
-    ):
-        remove_comma = True
-    elif tokens[first_brace].line == tokens[last_brace].line:
-        return None
-    else:
-        remove_comma = False
-
-    # determine the initial indentation
-    i = first_brace
-    while i >= 0 and tokens[i].name not in NEWLINES:
-        i -= 1
-
-    if i >= 0 and tokens[i + 1].name in INDENT_TOKENS:
-        initial_indent = len(tokens[i + 1].src)
-    else:
-        initial_indent = 0
-
-    return Fix(
-        (first_brace, last_brace),
-        multi_arg=multi_arg,
-        remove_comma=remove_comma,
-        initial_indent=initial_indent,
-    )
-
-
-def _find_call(
-        arg_offsets: Set[Offset],
-        i: int,
-        tokens: List[Token],
-) -> Optional[Fix]:
-    # When we get a `call` object, the ast refers to it as this:
-    #
-    #     func_name(arg, arg, arg)
-    #     ^ where ast points
-    #
-    # We care about the closing paren, in order to find it, we first walk
-    # until we find an argument.  When we find an argument, we know the outer
-    # paren we find is the function call paren
-    #
-    #     func_name(arg, arg, arg)
-    #              ^ outer paren
-    first_brace = None
-    paren_stack = []
-    for i in range(i, len(tokens)):
-        token = tokens[i]
-        if token.src == '(':
-            paren_stack.append(i)
-        # the ast lies to us about the beginning of parenthesized functions.
-        # See #3. (why we make sure there's something to pop here)
-        elif token.src == ')' and paren_stack:
-            paren_stack.pop()
-
-        if (token.line, token.utf8_byte_offset) in arg_offsets:
-            first_brace = paren_stack[0]
-            break
-    else:
-        raise AssertionError('Past end?')
-
-    return _find_simple(first_brace, tokens)
-
-
-def _find_tuple(i: int, tokens: List[Token]) -> Optional[Fix]:
-    # tuples are evil, we need to backtrack to find the opening paren
-    i -= 1
-    while tokens[i].name in NON_CODING_TOKENS:
-        i -= 1
-    # Sometimes tuples don't even have a paren!
-    # x = 1, 2, 3
-    if tokens[i].src != '(' and tokens[i].src != '[':
-        return None
-
-    return _find_simple(i, tokens)
-
-
-def _find_import(i: int, tokens: List[Token]) -> Optional[Fix]:
-    # progress forwards until we find either a `(` or a newline
-    for i in range(i, len(tokens)):
-        token = tokens[i]
-        if token.name == 'NEWLINE':
-            return None
-        elif token.name == 'OP' and token.src == '(':
-            return _find_simple(i, tokens)
-    else:
-        raise AssertionError('Past end?')
-
-
-def _fix_brace(
-        tokens: List[Token],
-        fix_data: Optional[Fix],
-        add_comma: bool,
-        remove_comma: bool,
-) -> None:
-    if fix_data is None:
-        return
-    first_brace, last_brace = fix_data.braces
-
-    # Figure out if either of the braces are "hugging"
-    hug_open = tokens[first_brace + 1].name not in NON_CODING_TOKENS
-    hug_close = tokens[last_brace - 1].name not in NON_CODING_TOKENS
-    if (
-            # Don't unhug single element things with a multi-line component
-            # inside.
-            not fix_data.multi_arg and
-            tokens[first_brace + 1].src in START_BRACES and
-            tokens[last_brace - 1].src in END_BRACES or
-            # Don't unhug when containing a single token (such as a triple
-            # quoted string).
-            first_brace + 2 == last_brace or
-            # don't unhug if it is a single line
-            fix_data.remove_comma
-    ):
-        hug_open = hug_close = False
-
-    # fix open hugging
-    if hug_open:
-        new_indent = fix_data.initial_indent + 4
-
-        tokens[first_brace + 1:first_brace + 1] = [
-            Token('NL', '\n'), Token(UNIMPORTANT_WS, ' ' * new_indent),
-        ]
-        last_brace += 2
-
-        # Adust indentation for the rest of the things
-        min_indent = None
-        indents = []
-        insert_indents = []
-        for i in range(first_brace + 3, last_brace):
-            if tokens[i - 1].name == 'NL' and tokens[i].name != 'NL':
-                if tokens[i].name != UNIMPORTANT_WS:
-                    min_indent = 0
-                    insert_indents.append(i)
-                else:
-                    if min_indent is None:
-                        min_indent = len(tokens[i].src)
-                    elif len(tokens[i].src) < min_indent:
-                        min_indent = len(tokens[i].src)
-                    indents.append(i)
-
-        if indents:
-            assert min_indent is not None
-            for i in indents:
-                oldlen = len(tokens[i].src)
-                newlen = oldlen - min_indent + new_indent
-                tokens[i] = tokens[i]._replace(src=' ' * newlen)
-        for i in reversed(insert_indents):
-            tokens.insert(i, Token(UNIMPORTANT_WS, ' ' * new_indent))
-            last_brace += 1
-
-    # fix close hugging
-    if hug_close:
-        tokens[last_brace:last_brace] = [
-            Token('NL', '\n'),
-            Token(UNIMPORTANT_WS, ' ' * fix_data.initial_indent),
-        ]
-        last_brace += 2
-
-    # From there, we can walk backwards and decide whether a comma is needed
-    i = last_brace - 1
-    while tokens[i].name in NON_CODING_TOKENS:
-        i -= 1
-
-    # If we're not a hugging paren, we can insert a comma
-    if add_comma and tokens[i].src != ',' and i + 1 != last_brace:
-        tokens.insert(i + 1, Token('OP', ','))
-
-    # Fix trailing brace to match leading indentation
-    back_1 = tokens[last_brace - 1]
-    back_2 = tokens[last_brace - 2]
-    if (
-            back_1.name == UNIMPORTANT_WS and
-            back_2.name == 'NL' and
-            len(back_1.src) != fix_data.initial_indent
-    ):
-        indent = fix_data.initial_indent * ' '
-        tokens[last_brace - 1] = back_1._replace(src=indent)
-
-    if fix_data.remove_comma:
-        start = last_brace
-        if tokens[start - 1].name == UNIMPORTANT_WS:
-            start -= 1
-        if remove_comma and tokens[start - 1].src == ',':
-            start -= 1
-        del tokens[start:last_brace]
-
-
-def _one_el_tuple(node: ast.expr) -> bool:
-    return isinstance(node, ast.Tuple) and len(node.elts) == 1
-
-
-def _changing_list(
-        lst: List[Token],
-) -> Generator[Tuple[int, Token], None, None]:
+def _changing_list(lst: List[Token]) -> Iterable[Tuple[int, Token]]:
     i = 0
     while i < len(lst):
         yield i, lst[i]
@@ -584,8 +45,8 @@ def _fix_src(contents_text: str, min_version: Tuple[int, ...]) -> str:
             callback(i, tokens, min_version)
 
         if token.src in START_BRACES:
-            _fix_brace(
-                tokens, _find_simple(i, tokens),
+            fix_brace(
+                tokens, find_simple(i, tokens),
                 add_comma=False,
                 remove_comma=False,
             )
